airflow_docker/requirements_airflow.txt
findspark
s3fs
apache-airflow-providers-apache-spark
apache-airflow-providers-docker
apache-airflow-providers-amazon